Documents/Investimento/Investimento/Investimento/dashboard/requirements-api.txt
flask==2.3.3
flask-cors==4.0.0
ccxt==4.1.13
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
